+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -134,7 +134,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-05-26T07:13:04.166Z")
+    updated: Optional[datetime] = parse_datetime("2025-05-26T10:36:44.128Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")