+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -83,5 +83,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-05-26T07:13:04.166Z")
+    updated: Optional[datetime] = parse_datetime("2025-05-26T10:36:44.128Z")
     r"""The date and time the resource was last updated"""