+++ packages/orq-rc/RELEASES.md
@@ -1499,3 +1499,13 @@ Based on:
 - [python v3.6.0-rc.35] packages/orq-rc
 ### Releases
 - [PyPI v3.6.0-rc.35] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.35 - packages/orq-rc
+
+## 2025-05-26 10:38:13
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.551.0 (2.610.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.6.0-rc.36] packages/orq-rc
+### Releases
+- [PyPI v3.6.0-rc.36] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.36 - packages/orq-rc